% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dt_code_from_csv.R
\name{generateDfCodeFromCsv}
\alias{generateDfCodeFromCsv}
\title{Generate data.frame code from CSV}
\usage{
generateDfCodeFromCsv(csv_file_path = NULL,
  object_type = c("data.frame", "data.table"))
}
\arguments{
\item{csv_file_path}{character, file path to the .csv file}

\item{object_type}{character, the object type for which to create the code.
Possible values are either "data.frame" (default) or "data.table"}
}
\value{
a character, which can be used to define a `data.frame` or `data.table`.
  Also prints the return value as a formatted string.
  This message is also copied to the user's clipboard.
}
\description{
Generate data.frame code from CSV
}
\examples{
\dontrun{
   csv_path <- tempfile("foo.csv")
   write.csv(
       data.frame(foo = c(1, 2, 3), bar = c("four", 5, "six")),
       csv_path, row.names = FALSE
   )
   generateDfCodeFromCsv(csv_path)

   unlink(csv_path)
}

}
